+++ packages/perseus/src/widget-ai-utils/interactive-graph/interactive-graph-ai-utils.ts
@@ -1,6 +1,6 @@
 import {UnreachableCaseError} from "@khanacademy/wonder-stuff-core";
 
-import type interactiveGraph from "../../widgets/interactive-graph";
+import type interactiveGraph from "../../widgets/interactive-graphs/interactive-graph";
 import type {UnsupportedWidgetPromptJSON} from "../unsupported-widget";
 import type {PerseusGraphType} from "@khanacademy/perseus-core";
 import type React from "react";