+++ packages/perseus/src/widgets/dropdown/__docs__/dropdown.stories.tsx
@@ -1,14 +1,13 @@
 import {generateTestPerseusItem} from "@khanacademy/perseus-core";
 
-import {ServerItemRendererWithDebugUI} from "../../../../../testing/server-item-renderer-with-debug-ui";
-
+import {ServerItemRendererWithDebugUI} from "../../../../../../testing/server-item-renderer-with-debug-ui";
 import {
     basicDropdown,
     dropdownWithEmptyPlaceholder,
     dropdownWithMath,
     dropdownWithVisibleLabel,
     inlineDropdownWithVisibleLabel,
-} from "./dropdown.testdata";
+} from "../dropdown.testdata";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
 
@@ -30,6 +29,7 @@ export default meta;
 
 type Story = StoryObj<typeof ServerItemRendererWithDebugUI>;
 
+// TODO(ivy): Find way not to include sorting of stories.
 export const BasicDropdown: Story = {
     args: {
         item: generateTestPerseusItem({question: basicDropdown}),