+++ packages/perseus/src/widgets/interactive-graphs/__docs__/interactive-graph-regression.stories.tsx
@@ -5,13 +5,12 @@ import {
 import {View} from "@khanacademy/wonder-blocks-core";
 import * as React from "react";
 
-import {ApiOptions} from "../../perseus-api";
-import Renderer from "../../renderer";
-import {mockStrings} from "../../strings";
-import UserInputManager from "../../user-input-manager";
-
-import {interactiveGraphQuestionBuilder} from "./interactive-graph-question-builder";
-import {sinusoidWithPiTicks} from "./interactive-graph.testdata";
+import {ApiOptions} from "../../../perseus-api";
+import Renderer from "../../../renderer";
+import {mockStrings} from "../../../strings";
+import UserInputManager from "../../../user-input-manager";
+import {interactiveGraphQuestionBuilder} from "../interactive-graph-question-builder";
+import {sinusoidWithPiTicks} from "../interactive-graph.testdata";
 
 import type {PerseusRenderer} from "@khanacademy/perseus-core";
 import type {Meta, StoryObj} from "@storybook/react-vite";